Deleted old semantic service and its components
